Update 1.4.3 features

+++ src/utilis/supportGen/writeFolder.ts
@@ -3,6 +3,11 @@ import path from 'path';
 
 function writeFolder(filename: string, data: string) {
     let currentPath = path.join(process.cwd(), filename);
+
+    if(fs.existsSync(currentPath)){
+        return false
+    }
+    
     fs.writeFileSync(currentPath, data);
     return true;
 }
